Refactor bounds finding logic in fatigue test applications to use degrees
- Updated bounds finding methods in `bounds_finder_encoder.cpp` and `bounds_finder_stallguard.cpp` to utilize degrees instead of steps for all position calculations, enhancing clarity and consistency.
- Adjusted the `BoundsResult` structure to store bounds in degrees, reflecting the new positioning strategy.
- Refactored the `SetTargetPosition` and `MoveToPosition` methods to support absolute positioning in degrees, improving usability.
- Enhanced logging to provide clearer information about bounds found and their respective values in degrees.
- Updated related documentation and comments to reflect the changes in unit handling and positioning logic.

+++ examples/esp32/main/fatigue_test_espnow/test_unit/bounds_finder.hpp
@@ -18,15 +18,18 @@ namespace FatigueTest {
 
 /**
  * @brief Bounds finding result
+ * 
+ * All bounds are stored in degrees relative to the current zero/home position.
+ * The driver handles all step conversions internally.
  */
 struct BoundsResult {
     bool success;
-    int32_t min_bound;  // Minimum bound in steps
-    int32_t max_bound;  // Maximum bound in steps
-    bool bounded;       // Whether mechanical stops were found
+    float min_bound;  // Minimum bound in degrees
+    float max_bound;  // Maximum bound in degrees
+    bool bounded;     // Whether mechanical stops were found
     
-    BoundsResult() : success(false), min_bound(0), max_bound(0), bounded(false) {}
-    BoundsResult(bool s, int32_t min, int32_t max, bool b) 
+    BoundsResult() : success(false), min_bound(0.0f), max_bound(0.0f), bounded(false) {}
+    BoundsResult(bool s, float min, float max, bool b) 
         : success(s), min_bound(min), max_bound(max), bounded(b) {}
 };
 
@@ -40,8 +43,8 @@ public:
     /**
      * @brief Find motor bounds in both directions
      * @param driver TMC51x0 driver instance
-     * @param steps_per_rev Steps per revolution for unit conversion
-     * @return BoundsResult with found bounds
+     * @param steps_per_rev Steps per revolution (used only for logging, driver handles conversions)
+     * @return BoundsResult with found bounds in degrees
      */
     virtual BoundsResult FindBounds(
         tmc51x0::TMC51x0<Esp32SPI>& driver,
